Rename plugin files to match their renamed Cedar exports
File-level follow-up to the function renames: rename useRedwoodX.ts ->
useCedarX.ts for the 7 internal-only plugins (and their test files) so
the file names match the exported function names. useRedwoodDirective.ts
is left as-is since it's out of scope for this PR (handled in #2317).

Also updates the few places that reference these names in comments/docs:
packages/api-server logFormatter test/README, tasks/ud-tests/udDev.test.mts,
and docs/implementation-docs/2026-03-26-cedarjs-project-overview.md.
