+++ packages/graphql-server/src/plugins/__tests__/useCedarGlobalContextSetter.test.ts
@@ -6,8 +6,8 @@ import type { GlobalContext } from '@cedarjs/context'
 import { context, setContext } from '@cedarjs/context'
 import { getAsyncStoreInstance } from '@cedarjs/context/dist/store'
 
-import { useCedarGlobalContextSetter } from '../useRedwoodGlobalContextSetter.js'
-import { useCedarPopulateContext } from '../useRedwoodPopulateContext.js'
+import { useCedarGlobalContextSetter } from '../useCedarGlobalContextSetter.js'
+import { useCedarPopulateContext } from '../useCedarPopulateContext.js'
 
 import { testSchema, testQuery } from './__fixtures__/common.js'
 import { createTestkit } from './__fixtures__/envelop-testing.js'
