+++ packages/graphql-server/src/plugins/index.ts
@@ -1,10 +1,10 @@
 export { useArmor } from './useArmor.js'
-export { useCedarAuthContext } from './useRedwoodAuthContext.js'
+export { useCedarAuthContext } from './useCedarAuthContext.js'
 export { useRedwoodDirective } from './useRedwoodDirective.js'
-export { useCedarError } from './useRedwoodError.js'
-export { useCedarGlobalContextSetter } from './useRedwoodGlobalContextSetter.js'
-export { useCedarLogger } from './useRedwoodLogger.js'
-export { useCedarTrustedDocuments } from './useRedwoodTrustedDocuments.js'
+export { useCedarError } from './useCedarError.js'
+export { useCedarGlobalContextSetter } from './useCedarGlobalContextSetter.js'
+export { useCedarLogger } from './useCedarLogger.js'
+export { useCedarTrustedDocuments } from './useCedarTrustedDocuments.js'
 
-export { useCedarPopulateContext } from './useRedwoodPopulateContext.js'
-export { useCedarOpenTelemetry } from './useRedwoodOpenTelemetry.js'
+export { useCedarPopulateContext } from './useCedarPopulateContext.js'
+export { useCedarOpenTelemetry } from './useCedarOpenTelemetry.js'
