+++ packages/graphql-server/src/plugins/useCedarGlobalContextSetter.ts
